+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 17.8.2015
+### ADDED
+- twitter-bootstrap-rails
+- bootstrap-datepicker-rails
+- select2-rails
+- jquery-turbolinks
+- google-webfonts-rails
+- nested-form
+- Layouts
+
 ## 16.8.2015
 ### ADDED
 - Devise with additional fields `username:string:uniq:index` and `name:string`